Refactor codebase
Completely refactor the existing codebase for improved structure and maintainability.
[skip gpt_engineer]
