+++ src/hooks/chat/utils/ChatError.ts
@@ -1,0 +1,6 @@
+export class ChatError extends Error {
+  constructor(message: string, public context: object) {
+    super(message)
+    this.name = 'ChatError'
+  }
+}
